pyproject.toml
[tool.poetry]
name = "ice"
version = "0.1.0"
description = ""
authors = ["Ought <team@ought.org>"]

[tool.poetry.dependencies]
python = "3.10.4"
defopt = "^6.3.0"
questionary = "^1.10.0"
rich = "^12.4.1"
pre-commit = "^2.19.0"
numpy = "^1.22.3"
rouge-metric = "^1.0.1"
fastapi = "^0.78.0"
uvicorn = { extras = ["standard"], version = "^0.17.6" }
ulid-py = "^1.1.0"
pydantic = "^1.9.1"
httpx = "^0.23.0"
sqlmodel = "^0.0.6"
sqlalchemy2-stubs = "^0.0.2-alpha.22"
nest-asyncio = "^1.5.5"
structlog = "^21.5.0"
alembic = "^1.7.7"
asyncpg = "^0.25.0"
transformers = "^4.19.2"
pandas = "^1.4.2"
networkx = "^2.8.2"
pyee = "^9.0.4"
Faker = "^13.12.1"
nltk = "^3.7"
tenacity = "^8.0.1"
thefuzz = "^0.19.0"
python-Levenshtein = "^0.12.2"
sentencepiece = "^0.1.96"
protobuf = "^3.20.1"
plotly = "^5.9.0"
stqdm = "^0.0.4"
scikit-learn = "^1.1.1"
tqdm = "^4.64.0"
torch = [
    { url = "https://download.pytorch.org/whl/cu116/torch-1.12.0%2Bcu116-cp310-cp310-linux_x86_64.whl", markers = "platform_machine == 'x86_64'" },
    { url = "https://download.pytorch.org/whl/cpu/torch-1.12.0-cp310-cp310-manylinux2014_aarch64.whl", markers = "platform_machine == 'aarch64'" }
]
websockets = "^10.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
watchdog = "^2.1.6"
pylint = "v3.0.0a3"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
mypy = "v0.971"
pre-commit = "^2.17.0"
autoflake = "^1.4"
flake8 = { version = "^4.0.1", optional = true }
streamlit = "^1.10.0"
scipy = "^1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
